[FE] Enhance UI
