+++ server/src/common/pagy/pagy.constant.ts
@@ -1,4 +1,4 @@
 export const PagyConstant = {
   DefaultPage: 1,
-  DefaultCount: 10,
+  DefaultCount: 15,
 };
